add UpdateGameAchievementUnlockMediansAction
